Fix cluster ssh to connect to private ip if required
When a machine is inside EC2 VPC and it doesn't have public ip
then the cluster ssh need to know to connect to the private ip
instead of the public hostname.

This patchs adds new connect_hostname() function which tries
to return public hostname, but will fall back to private ip.
This function is then used in cluster ssh